backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg[binary]==3.1.18
pydantic-settings==2.2.1
alembic==1.13.1
redis==5.0.1
python-dotenv==1.0.1
boto3==1.34.69
httpx==0.27.0
PyJWT==2.9.0
